requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
python-dotenv
Session
flask_session
sqlalchemy